+++ SourceCode/Data/GroupMember.cs
@@ -10,7 +10,8 @@ public partial class GroupMember
     public int PersonId { get; set; }
     public bool IsGroupAdministrator { get; set; }
     public bool IsDataAdministrator { get; set; }
-    public bool MayBorrowModules { get; set; }
+    public bool MayBorrowGroupsModules { get; set; }
+    public bool MemberMayBorrowMyModules { get; set; }
 
     public virtual Group Group { get; set; }
     public virtual Person Person { get; set; }